// FCUBEAPP/src/app/models/driversalarydetailmodel.ts
export class Driversalarydetailmodel {
    vehicleMasterId: string = "";
    driverMasterId: string = "";
    fromDt: string = "";
    toDt: string = "";
 
    salaryDays: string = "";
    salaryAmt: string = "";
    poolAmt: string = "";
    lastTripBal: string = "";
  //  lastTripBal: string = "";
    netPayable: string = "";
   // detRemarks: string = "";
   
   
}